requirements.txt
# Core framework
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0

# Vector database and embeddings
qdrant-client>=1.7.0
sentence-transformers>=2.2.0

# Web interface
streamlit>=1.28.0

# Datasets
datasets>=2.14.0
huggingface-hub>=0.19.0

# Web search (free)
duckduckgo-search>=3.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
uuid